requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
websockets>=11.0.3
telegram-send>=0.25
discord.py>=2.0.0
discord-webhook>=1.1.0
python-telegram-bot>=20.6
textblob>=0.17.1
vader-sentiment>=3.3.2
python-dateutil>=2.8.2
aiohttp>=3.8.5
asyncio>=3.4.3